+++ CHANGELOG.md
@@ -4,7 +4,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This change log adheres to standards from [Keep a CHANGELOG](http://keepachangelog.com).
 
 ## [Unreleased]
-- Peer dependencies update, inc. eslint to v1.7.3 and eslint-plugin-react to v3.6.3
+### Added
+- `arrow-body-style` new rule from eslint v1.8.0
+- `no-arrow-condition` new rule from eslint v1.8.0
+
+### Changed
+- `no-multiple-empty-lines`: new option `maxEOF` set
+- Peer dependencies update, inc. eslint to v1.8.0 and eslint-plugin-react to v3.6.3
 
 
 ## [1.7.0] - 2015-10-18